HAPP-195: Rename RecordCommentsAdapter
